pyproject.toml
[tool.poetry]
name = "test-service"
version = "0.1.0"
description = ""
authors = ["Volodymyr Vereshchak <volodymyr.vereshchak@gmail.com>"]
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["all"], version = "^0.89.0"}
dropbox = "^11.36.0"
redis = "^4.4.1"
python-dotenv = "^0.21.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
